Combined the Project tests and fixed the formatting.
